requirements.txt
streamlit==1.24.0
requests==2.31.0
matplotlib==3.7.1
pandas==2.0.3
plotly==5.15.0
python-dotenv==1.0.0